Add gconnect, gdisconnect. Change login features. Stopped using login-flask and instead started using session.

+++ Project3/vagrant/catalog/project.py
@@ -25,6 +25,12 @@ from oauth2client.client import flow_from_clientsecrets
 from oauth2client.client import FlowExchangeError
 import httplib2
 
+# Google login settings
+CLIENT_ID = json.loads(
+    open('client_secrets.json', 'r').read())['web']['client_id']
+APPLICATION_NAME = "My Goals App"
+
+# File upload settings
 UPLOAD_FOLDER = 'static/uploads'
 ALLOWED_EXTENSIONS = set(['png', 'jpg', 'jpeg', 'gif'])
 
@@ -38,6 +44,7 @@ login_manager = LoginManager()
 login_manager.init_app(app)
 login_manager.login_view = "showLogin"
 
+#flask-login method
 @login_manager.user_loader
 def load_user(user_id):
 	return session.query(User).filter_by(id = user_id).one()
@@ -51,6 +58,26 @@ Base.metadata.bind = engine
 DBSession = sessionmaker(bind=engine)
 session = DBSession()
 
+# Local permission system methods
+def createUser(login_session):
+  newUser = User(username = login_session['username'], email = login_session['email'], picture = login_session['picture'])
+  session.add(newUser)
+  session.commit()
+  user = session.query(User).filter_by(email = login_session['email']).one()
+  return user.id
+
+def getUserID(email):
+  try:
+    user = session.query(User).filter_by(email = email).one()
+    return user.id
+  except:
+    return None
+
+def getUserInfo(user_id):
+  user = session.query(User).filter_by(id = user_id).one()
+  return user
+
+# file upload method
 def allowed_file(filename):
 	''' Check if the file is allowed '''
 	return '.' in filename and filename.rsplit('.', 1)[1] in ALLOWED_EXTENSIONS
@@ -73,21 +100,145 @@ def showLogin():
 	state = ''.join(random.choice(string.ascii_uppercase + string.digits) \
 	 for x in xrange(32))
   	login_session['state'] = state
-  	if request.method == 'POST':
-  		if request.form['id']:
-  			user_id = request.form['id']
-  		user = session.query(User).filter_by(id = user_id).one()
-  		login_user(user)
-  		return redirect(url_for('showIndex'))
-  	else:
-	  	#RENDER THE LOGIN TEMPLATE
-	  	return render_template('login.html', STATE=state)
+  	return render_template('login.html', STATE=state)
+  	# if request.method == 'POST':
+  	# 	if request.form['id']:
+  	# 		user_id = request.form['id']
+  	# 	user = session.query(User).filter_by(id = user_id).one()
+  	# 	login_user(user)
+  	# 	return redirect(url_for('showIndex'))
+  	# else:
+	  # 	#RENDER THE LOGIN TEMPLATE
+	  # 	return render_template('login.html', STATE=state)
 
 @app.route("/logout")
-@login_required
+#@login_required
 def logout():
-    logout_user()
-    return redirect(url_for('showIndex'))
+	#Check if user is logged in
+	if 'username' not in login_session:
+		return redirect('/login')
+	gdisconnect()
+	return redirect(url_for('showIndex'))
+    #logout_user()
+    #return redirect(url_for('showIndex'))
+
+@app.route('/gconnect', methods=['POST'])
+def gconnect():
+  # Validate state token
+  if request.args.get('state') != login_session['state']:
+	  response = make_response(json.dumps('Invalid state parameter.'), 401)
+	  response.headers['Content-Type'] = 'application/json'
+	  return response
+  # Obtain authorization code
+  code = request.data
+
+  try:
+      # Upgrade the authorization code into a credentials object
+      oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
+      oauth_flow.redirect_uri = 'postmessage'
+      credentials = oauth_flow.step2_exchange(code)
+  except FlowExchangeError:
+      response = make_response(
+          json.dumps('Failed to upgrade the authorization code.'), 401)
+      response.headers['Content-Type'] = 'application/json'
+      return response
+
+  # Check that the access token is valid.
+  access_token = credentials.access_token
+  url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
+         % access_token)
+  h = httplib2.Http()
+  result = json.loads(h.request(url, 'GET')[1])
+  # If there was an error in the access token info, abort.
+  if result.get('error') is not None:
+      response = make_response(json.dumps(result.get('error')), 500)
+      response.headers['Content-Type'] = 'application/json'
+
+  # Verify that the access token is used for the intended user.
+  gplus_id = credentials.id_token['sub']
+  if result['user_id'] != gplus_id:
+      response = make_response(
+          json.dumps("Token's user ID doesn't match given user ID."), 401)
+      response.headers['Content-Type'] = 'application/json'
+      return response
+
+  # Verify that the access token is valid for this app.
+  if result['issued_to'] != CLIENT_ID:
+      response = make_response(
+          json.dumps("Token's client ID does not match app's."), 401)
+      print "Token's client ID does not match app's."
+      response.headers['Content-Type'] = 'application/json'
+      return response
+
+  stored_credentials = login_session.get('credentials')
+  stored_gplus_id = login_session.get('gplus_id')
+  if stored_credentials is not None and gplus_id == stored_gplus_id:
+      response = make_response(json.dumps('Current user is already connected.'),
+                               200)
+      response.headers['Content-Type'] = 'application/json'
+      return response
+
+  # Store the access token in the session for later use.
+  login_session['credentials'] = credentials
+  login_session['gplus_id'] = gplus_id
+
+  # Get user info
+  userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
+  params = {'access_token': credentials.access_token, 'alt': 'json'}
+  answer = requests.get(userinfo_url, params=params)
+  data = answer.json()
+
+  login_session['username'] = data['name']
+  login_session['picture'] = data['picture']
+  login_session['email'] = data['email']
+
+  #see if user exists if it doesn't make a new one
+  user_id = getUserID(login_session['email'])
+  if not user_id:
+    user_id = createUser(login_session)
+  login_session['user_id'] = user_id
+
+  output = ''
+  output += '<h1>Welcome, '
+  output += login_session['username']
+  output += '!</h1>'
+  output += '<img src="'
+  output += login_session['picture']
+  output += ' " style = "width: 300px; height: 300px;border-radius: 150px; \
+  -webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+  flash("you are now logged in as %s" % login_session['username'])
+  print "done!"
+  return output
+
+@app.route('/gdisconnect')
+def gdisconnect():
+    # Only disconnect a connected user.
+    credentials = login_session.get('credentials')
+    if credentials is None:
+        response = make_response(
+            json.dumps('Current user not connected.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    access_token = credentials.access_token
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % access_token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[0]
+    if result['status'] == '200':
+        del login_session['gplus_id']
+        del login_session['credentials']
+        del login_session['username']
+        del login_session['email']
+        del login_session['picture']
+
+        response = make_response(json.dumps('Successfully disconnected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    else:
+    	response = make_response(
+            json.dumps('Failed to revoke token for given user.', 400))
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
 
 
 @app.route('/<int:goal_id>/goal/', methods=['GET', 'POST'])
@@ -104,10 +255,10 @@ def showGoal(goal_id):
 	if request.method == 'POST':
 
 		#current_user(signed in user) <--- flask-login method
-		if current_user.is_authenticated:
+		if 'username' in login_session:
 			newComment = Comments(content=request.form['content'],
 							timestamp=datetime.datetime.utcnow(),
-							user_id = current_user.id,
+							user_id = login_session['user_id'],
 							goal_id = userGoal.Goal.id)
 
 			session.add(newComment)
@@ -129,6 +280,16 @@ def showProfile(user_id):
 @app.route('/user/<int:user_id>/goal/new/', methods=['GET','POST'])
 def newGoal(user_id):
 	''' Handler function for 'Create Goal' page '''
+
+	user = session.query(User).filter_by(id = user_id).one()
+	creator = getUserInfo(user.id)
+	#Check if user is logged in
+	if 'username' not in login_session:
+		return redirect('/login')
+
+	if creator.id != login_session['user_id']:
+		return redirect('/')
+
 	if request.method == 'POST':
 		file = request.files['file']
 
@@ -185,9 +346,19 @@ def newGoal(user_id):
  methods=['GET', 'POST'])
 def editGoal(user_id, goal_id):
 	''' Handler function for a 'Edit Goal' page '''
-	#return 'This lets a user edit a goal'.
 
 	editedGoal = session.query(Goal).filter_by(id = goal_id).one()
+	creator = getUserInfo(editedGoal.user_id)
+
+	#Check if user is logged in
+	if 'username' not in login_session:
+		return redirect('/login')
+
+	if creator.id != login_session['user_id']:
+		return redirect('/')
+
+	#return 'This lets a user edit a goal'.
+
 	if request.method == 'POST':
 		''' File Handler '''
 		file = request.files['file']
@@ -249,8 +420,18 @@ def editGoal(user_id, goal_id):
 	methods=['GET', 'POST'])
 def deleteGoal(user_id, goal_id):
 	''' Handler function for a 'Delete Goal' page '''
-	#return 'This lets a user delete a goal'
+
 	goalToDelete = session.query(Goal).filter_by(id = goal_id).one()
+	creator = getUserInfo(goalToDelete.user_id)
+
+	if 'username' not in login_session:
+		return redirect('/login')
+
+	if creator.id != login_session['user_id']:
+		return redirect('/')
+
+	#return 'This lets a user delete a goal'
+	
 	if request.method == 'POST':
 		session.delete(goalToDelete)
 		session.commit()
@@ -258,9 +439,17 @@ def deleteGoal(user_id, goal_id):
 	else:
 		return render_template('deleteGoal.html', goal = goalToDelete)
 
-@app.route('/<int:goal_id>/goal/complete/')
-def completeGoal(goal_id):
+@app.route('/user/<int:user_id>/<int:goal_id>/goal/complete/')
+def completeGoal(user_id, goal_id):
 	''' Handler function for a 'Completeing Goal' page '''
+	user = session.query(User).filter_by(id = user_id).one()
+	creator = getUserInfo(user.id)
+
+	if 'username' not in login_session:
+		return redirect('/login')
+
+	if creator.id != login_session['user_id']:
+		return redirect('/')	
 	#return 'This lets a user mark a goal complete'
 	return render_template('completeGoal.html')
 
@@ -268,7 +457,16 @@ def completeGoal(goal_id):
 	methods=['GET', 'POST'])
 def editProfile(user_id):
 	''' Handler function for a specific user to edit their profile '''
+
 	editedUser = session.query(User).filter_by(id = user_id).one()
+	creator = getUserInfo(editedUser.id)
+
+	#Check if user is logged in
+	if 'username' not in login_session:
+		return redirect('/login')
+
+	if creator.id != login_session['user_id']:
+		return redirect('/')
 
 	if request.method == 'POST':
 		file = request.files['file']
@@ -287,13 +485,13 @@ def editProfile(user_id):
 			editedUser.picture = os.path.join('/'+ app.config['UPLOAD_FOLDER'],
 			 f_name)
 
-			if request.form['description']:
-				editedUser.desc = request.form['description']
+		if request.form['description']:
+			editedUser.desc = request.form['description']
 
-			session.add(editedUser)
-			session.commit()
-			return redirect(url_for('showProfile', 
-				user_id = editedUser.id))
+		session.add(editedUser)
+		session.commit()
+		return redirect(url_for('showProfile', 
+			user_id = editedUser.id))
 	else:
 	#return 'This allows a user (%s) to edit their profile' %username
 		return render_template('editProfile.html', user = editedUser)
